Initial commit of User Guide WIP.
Sweeping structural changes and lots of added screenshots.  There is a lot of initial work in this commit and needs to be saved, but its not in any 'publishable' state.  This is mostly a 'backup commit' to not lose the large amount of work done thus far, but much still remains.